requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
defusedxml==0.8.0rc2
discord.py==2.5.2
Django==5.0.6
django-allauth==0.63.3
django-cors-headers==4.4.0
djangorestframework==3.15.2
frozenlist==1.5.0
idna==3.7
jwt==1.3.1
multidict==6.2.0
oauthlib==3.2.2
propcache==0.3.1
pycparser==2.22
PyJWT==2.8.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
yarl==1.18.3
